Move KeyedAccount out of solana-program.  Native programs are not supported by solana-program
