+++ programs/bpf/tests/programs.rs
@@ -18,12 +18,13 @@ use solana_runtime::{
     },
 };
 use solana_sdk::{
-    account::{Account, KeyedAccount},
+    account::Account,
     bpf_loader, bpf_loader_deprecated,
     client::SyncClient,
     clock::{DEFAULT_SLOTS_PER_EPOCH, MAX_PROCESSING_AGE},
     entrypoint::{MAX_PERMITTED_DATA_INCREASE, SUCCESS},
     instruction::{AccountMeta, CompiledInstruction, Instruction, InstructionError},
+    keyed_account::KeyedAccount,
     message::Message,
     pubkey::Pubkey,
     signature::{Keypair, Signer},
