+++ programs/failure/src/lib.rs
@@ -1,5 +1,5 @@
-use solana_sdk::account::KeyedAccount;
 use solana_sdk::instruction::InstructionError;
+use solana_sdk::keyed_account::KeyedAccount;
 use solana_sdk::pubkey::Pubkey;
 
 solana_sdk::declare_program!(
