+++ programs/vest/src/vest_processor.rs
@@ -7,8 +7,9 @@ use chrono::prelude::*;
 use solana_config_program::date_instruction::DateConfig;
 use solana_config_program::get_config_data;
 use solana_sdk::{
-    account::{next_keyed_account, Account, KeyedAccount},
+    account::Account,
     instruction::InstructionError,
+    keyed_account::{next_keyed_account, KeyedAccount},
     program_utils::limited_deserialize,
     pubkey::Pubkey,
 };
