+++ runtime/src/message_processor.rs
@@ -12,9 +12,10 @@ use crate::{
 use log::*;
 use serde::{Deserialize, Serialize};
 use solana_sdk::{
-    account::{create_keyed_readonly_accounts, Account, KeyedAccount},
+    account::Account,
     clock::Epoch,
     instruction::{CompiledInstruction, Instruction, InstructionError},
+    keyed_account::{create_keyed_readonly_accounts, KeyedAccount},
     message::Message,
     native_loader,
     pubkey::Pubkey,
