+++ sdk/src/lib.rs
@@ -18,8 +18,10 @@ pub mod genesis_config;
 pub mod hard_forks;
 pub mod hash;
 pub mod inflation;
+pub mod keyed_account;
 pub mod log;
 pub mod native_loader;
+pub mod nonce_keyed_account;
 pub mod packet;
 pub mod poh_config;
 pub mod program_utils;
